pyproject.toml
[tool.poetry]
name = "schoolmanagement"
version = "0.1.0"
description = ""
authors = ["avijeetjha77 <avijeetjha@rirabh.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"
django-rest-passwordreset = "1.3.0"
django-redis = "^5.4.0"
django-split-settings = "^1.3.0"
pyyaml = "^6.0.1"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
